+++ easy_task/lib/src/screens/task.create.screen.dart
@@ -24,7 +24,7 @@ class _TaskCreateScreenState extends State<TaskCreateScreen> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: const Text('Create Task'),
+        title: Text('Create Task'.t),
         actions: const [],
       ),
       body: Padding(
@@ -39,11 +39,13 @@ class _TaskCreateScreenState extends State<TaskCreateScreen> {
             ),
             TextField(
               controller: _titleController,
-              decoration: const InputDecoration(labelText: 'Title'),
+              decoration: InputDecoration(labelText: 'Title'.t),
             ),
             TextField(
               controller: _descriptionController,
-              decoration: const InputDecoration(labelText: 'Description'),
+              decoration: InputDecoration(labelText: 'Task Description'.t),
+              minLines: 3,
+              maxLines: 8,
             ),
             const SizedBox(height: 20),
             UploadForm(
@@ -53,7 +55,7 @@ class _TaskCreateScreenState extends State<TaskCreateScreen> {
               button: ElevatedButton(
                 onPressed: () async {
                   if (_titleController.text.isEmpty) {
-                    toast(context: context, message: Text('enter a title'.t));
+                    toast(context: context, message: Text('input a title'.t));
                     return;
                   }
                   final ref = await Task.create(
@@ -73,7 +75,7 @@ class _TaskCreateScreenState extends State<TaskCreateScreen> {
                     }
                   }
                 },
-                child: const Text('Create Task'),
+                child: Text('Create Task'.t),
               ),
             ),
           ],