+++ easychat/lib/easychat.dart
@@ -1,6 +1,22 @@
+// Models =====================================================
 export './src/chat.message.dart';
 export './src/chat.room.dart';
+export './src/chat.room.user.dart';
+
+// Service ====================================================
 export './src/chat.service.dart';
+
+// Widgets ====================================================
+export './src/widgets/chat.bubble.dart';
+export './src/widgets/chat.messages.list_view.dart';
+export './src/widgets/chat.room.input_box.dart';
+export './src/widgets/chat.room.invitation.list_tile.dart';
+export './src/widgets/chat.room.list_view.dart';
+export './src/widgets/chat.room.list_tile.dart';
+
+// Screens =====================================================
+export './src/widgets/screens/chat.room.edit.screen.dart';
+export './src/widgets/screens/chat.room.list.screen.dart';
 export './src/widgets/screens/chat.room.screen.dart';
-export 'src/widgets/screens/chat.room.edit.screen.dart';
-export 'src/widgets/screens/chat.room.list.screen.dart';
+export './src/widgets/screens/received.chat.room.invite_list.screen.dart';
+export './src/widgets/screens/rejected.chat.room.invite_list.screen.dart';